Move the remaining `pipeline` into `genfile`
