+++ libs/dm-localization/package.json
@@ -1,5 +1,6 @@
 {
   "name": "@genshin-optimizer/dm-localization",
   "version": "0.0.1",
-  "type": "commonjs"
+  "type": "commonjs",
+  "executors": "./executors.json"
 }
