+++ libs/dm-localization/src/executors/gen-locale/executor.spec.ts
@@ -1,0 +1,8 @@
+import executor from './executor'
+
+describe.skip('GenLocale Executor', () => {
+  it('can run', async () => {
+    const output = await executor({})
+    expect(output.success).toBe(true)
+  })
+})
