+++ libs/dm-localization/src/executors/gen-locale/lib/Data/Characters.ts
