+++ libs/dm-localization/src/executors/gen-locale/lib/common.ts
