+++ libs/dm-localization/src/executors/gen-locale/schema.d.ts
@@ -1,0 +1,1 @@
+export interface GenLocaleExecutorSchema {} // eslint-disable-line
