+++ libs/dm-localization/src/executors/gen-locale/schema.json
@@ -1,0 +1,9 @@
+{
+  "$schema": "http://json-schema.org/schema",
+  "version": 2,
+  "title": "GenLocale executor",
+  "description": "",
+  "type": "object",
+  "properties": {},
+  "required": []
+}
