+++ libs/localization/project.json
@@ -4,13 +4,15 @@
   "sourceRoot": "libs/localization/src",
   "projectType": "library",
   "targets": {
-    "pipeline": {
-      "command": "ts-node libs/localization/src/loadTrans.ts",
-      "inputs": ["production"]
-    },
-    "mainGen": {
-      "command": "ts-node libs/localization/src/mainGen.ts",
-      "inputs": ["production"]
+    "genfile": {
+      "executor": "nx:run-commands",
+      "options": {
+        "commands": [
+          "ts-node libs/localization/src/loadTrans.ts",
+          "ts-node libs/localization/src/mainGen.ts"
+        ]
+      },
+      "outputs": ["{projectRoot}/assets/locales", "{projectRoot}/main_gen.json"]
     },
     "lint": {}
   },
